[Ray] Move some tests out of test_ray.py (#3109)
* move some tests to ray_cluster

* rename test_ray_cluster to test_ray_cluster_standalone

(cherry picked from commit 379c9fbb823456de72fc808f9e72f84e5ccf2882)
